Fix HDP2 Tokens verification
HDP3 is able to read various token that have been migrated to protobuf in the old format, but during sasl verification steps, the receiver (HDPv3.3) verifies the token was correctly encoded by the sender (HPDv2.6). To do so it serializes the token from the extracted fields, using the new format (proto).
Due to that, the encoded token by the receiver does not match the one encoded by the sender and the communication fails with a security issue.

This commit solves the issue by:
  - setting a flag to remember if a token was extracted from an old binary format
  - if the flag is set, when the token is serialized, it is serialized using the old format

The commit only covers proto token that can be sent by HDP2