LAKE-15323 Hadoop 3 can read FileStatus serialized with Hadoop2 without DataInput copy (#20)
Previous attempt to fix the issue was based on exception, and thus required DataInput copy. This is not compatible with ConsolidatorInput criteo class that provides a longer data stream containing many FileStatuses serialized.

This fix detects hadoop 2 vs hadoop 3 serialization based on the first byte read. Basically, with hadoop 3, this first byte is supposed to equal 0, and should always be set with hadoop 2. Details are present in the code itself.

Co-authored-by: William Montaz <w.montaz@criteo.com>