Fix FileStatus deserialisation for hadoop 3
The previous commit does not nicely recreate the integer from 4 bytes, that is because bytes are signed but to reconstruct an int we want to get all bits of the byte without taking care of the sign.
To convert byte to unsigned int, one must use a bitmask with & operator.