Preventing class cast errors on ArrayWritable.toArray()
The method uses Arrays.copy which will not use the valueClass. The previous version of the code (before HADOOP-16678) was explicitly creating an array of the desired type and populating it. It was a bit less efficient than Arrays.copy but other approaches to try to infer the class to Arrays.copy method seems a bit convoluted and are hard to get right without brute casting (one needs to create a Class object representing an array of valueClass, everything is made by reflection and gets hard to get right).