[CRITEO] Add a property preventing scheduling on non exclusive partitions
During the simulation of hadoop on 2 datacenters, we had to use a non-exclusive partition to smoothly migrate jobs to their simulated placement group. However, scheduling is always authorized on non-exclusive partitions, even if the max capacity is 0. This property prevents any such scheduling to happen.

When no more testing is required, we can remove this change.