[CRITEO] Allow datanode bandwidth update to update any throttler
To do this, we use some bits of the long value transmitted to datanode to chose the throttler. This allows very few changes in the code (vs having a dedicated field for that purpose).

The bandwidth value is a long that contains the type of throttler in the 2nd, 3rd and 4th highest bit
(to avoid using the 1st highest bit and require inputting negative values)
  * TRANSFER throttler is 2
  * WRITE throttler is 3
  * any other value acts as the default -> BALANCE throttler

Example, for 500MB/sec and throttler write, the long value is 3458764514344828928
which is 0011000000000000000000000000000000011111010000000000000000000000
  * the throttler is represented by 0011 equals 3 -> WRITE throttler
  * the bandwidth is 11111010000000000000000000000 -> 524288000 = 500 * 1024 * 1024