[CRITEO] Remove final modifier on FileStatus.isDir() for DataDisco
DataDisco overrides this method as it is necessary to do so with Hadoop 2 base code. The final modifier was added to prevent FileStatus subclasses to implement this deprecated method and it actually calls the isDirectory method directly.

Once hadoop 3 migration is done, DataDisco can stop overriding this method and this commit can be reverted.