[CRITEO] Add a ReplicaCachingGetSpaceUsedWithTrash class to calculate datanode disk usage
ReplicaCachingGetSpaceUsed does not include trash size and that can be harmfull during rolling upgrade as blocks are moved to trash folder and are not deleted. This class aims at removing that. When trash is not activated or the folder does not exists, it has no effect.

(cherry picked from commit 802b168803b41e66b21ddacd946cc3da58ede8cc)