[CRITEO] Send signal to session instead of process group
The current code assumes that the process hierarchy inside a container
is composed of a single process group. But that's not always the case,
for example when a process calls setpgid(). This creates an issue when
sending signals to the processes inside a container.

This changes the signal_container_as_user() function to use the session
instead of the process group to signal all the processes in a container.
This is the most pragmatic way of achieving the desired results without
changing the container-executor implementation too much. For now, it
does the job.

Keep in mind that launch_container_as_user calls setsid() so we are
assured to be in a new session.

(cherry picked from commit e940b1f77fda2fe63fce60f8051ef021e3c434b4)