[CRITEO] Fix cgroup path computation when docker is enabled When docker is enabled, /proc/pid/cgroup shows path as following:
7:cpu,cpuacct:/hadoop-yarn/container_e72_1712241417859_1971906_01_000958/de5b77f2cdf6b5e87de9c23da5bc07de0d69f010af503dcd15832a3389c31565

If we only capture the filename of the path (previous code), we end up building a cgroup path that does not exist. Instead we propose to match the container id to determine the final path. This could break if cgroup paths created by yarn woudl differ but it is very unlikely. If we don't find the container_id in the path, we log an ERROR which would be captured easily in logs.

(cherry picked from commit 16f5d29f6c06f0c290b86471252dc319ab2fa1a3)