[CRITEO] All included NM appear as NEW at startup
Then if they reconnect they will be running, and if not they eventually become LOST because the NMLivelinessMonitor will fire an EXPIRE event.

If LOST state was the first state of any nodemanager, it would trigger an immediate decommission if the nodes are under a maintenance.
Thus instead of directly calling the transition to LOST, we let the NMLivelinessMonitor do the job. This way we have the same timeout to consider NM LOST.

Putting the NM in the active NM set at startup causes many issues because a lot of components try to perform operations on this active set, thus it is easier to put them in the inactive set.
However, doing so, the EXPIRE event triggered by NMLivelinessMonitor cannot reach the RMNode because the dispatcher only transmit events to RMNode in the active set. A small change to the dispatcher helps to cope with this.

When upgrading to 3.5, check the behavior of https://issues.apache.org/jira/browse/YARN-11730