[CRITEO] Limit concurrency on xferService (#79) This commit adds a semaphore to block limit the number of threads that will actually perform DataTransfer, configurable via dfs.datanode.replication.non-ec.streams.concurrent.count with default to 0 (no limit)
An additional metric XmitsConcurrent has been added on the JMXBean of the datanode.

https://criteo.atlassian.net/wiki/spaces/HAD/pages/3418621762/Optimizing+HDFS+node+decommission+recovery

Another approach would have consisted in capping the ThreadPool of the xferService and I would have honestly done that as a preference, however, for a correct value of xmitsInProgress and thus a proper feeding of the datanode's ReplicationWork queue, it would have require to move the xmitsInProgress counter from outside of the DataTransfer task to when the task is submitted to the ThreadPool. I considered this approach to be a bit harder to maintain on the long run (one could forget to remove xmitsInProgress from the original code of a new branch and the counter would be incremented twice and quickly drift). Side note: capping the ThreadPool is really very close to what has been done with ErasureCoded work (because it could not be as much parallel as the DataTransfer, and there was problem linked to bad xmitsInProgress value, which was fixed by pretty much the same approach consisting in incrementing the counter at the moment the task is submitted.... see https://issues.apache.org/jira/browse/HDFS-12044)

(cherry picked from commit c0d35d0fd6df1809bc3c262e4f9a834926a1ce93)